post information navigation added from community posts

+++ screens/home/CommunityScreen.js
@@ -1,24 +1,19 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View, Text, StyleSheet, TouchableOpacity, ScrollView } from 'react-native';
 import { collection, query, getDocs } from 'firebase/firestore';
 import { FIREBASE_DB } from '../../config/firebase';
-import {FIREBASE_AUTH} from '../../config/firebase';
-
-import { ScrollView } from 'react-native-gesture-handler';
-import { TouchableOpacity } from 'react-native-gesture-handler';
+import { FIREBASE_AUTH } from '../../config/firebase';
 import { deleteDoc, doc } from 'firebase/firestore';
-
-import { Button } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import { ROUTES } from '../../constants';
-
 import { orderBy } from 'firebase/firestore';
+
 const CommunityScreen = () => {
   const navigation = useNavigation();
   const [sessions, setSessions] = useState([]);
 
   const fetchSessions = async () => {
-    const q = query(collection(FIREBASE_DB, 'community-chat'), orderBy('postCreatedDateTime'));
+    const q = query(collection(FIREBASE_DB, 'community-chat'), orderBy('postCreatedDateTime', 'desc'));
     const querySnapshot = await getDocs(q);
     const sessionData = [];
     querySnapshot.forEach((doc) => {
@@ -32,14 +27,12 @@ const CommunityScreen = () => {
     fetchSessions();
   }, []);
 
-  const reversedSessions = [...sessions].reverse();//reverse the list, so that msot recent posts are displayed on top
-
   const deleteSession = async (postId, userId) => {
     try {
-      // if (FIREBASE_AUTH.currentUser.uid !== userId) {  //this will be in the form of button 
-      //   alert('You can only delete your own post');
-      //   return;
-      // }
+      if (FIREBASE_AUTH.currentUser.uid !== userId) {
+        alert('You can only delete your own post');
+        return;
+      }
       await deleteDoc(doc(FIREBASE_DB, 'community-chat', postId));
       await fetchSessions();
     } catch (error) {
@@ -49,56 +42,43 @@ const CommunityScreen = () => {
 
   return (
     <ScrollView style={styles.container}>
-        <TouchableOpacity
-                style={styles.addButton}
-                onPress={() => navigation.navigate(ROUTES.ADD_POST_SCREEN)} 
-              >
-                <Text style={styles.deleteButtonText}>Add post</Text>
-      </TouchableOpacity>
+      <TouchableOpacity style={styles.addButton} onPress={() => navigation.navigate(ROUTES.ADD_POST_SCREEN)}>
+        <Text style={styles.deleteButtonText}>Add post</Text>
+      </TouchableOpacity> 
       <View style={styles.container}>
-      {reversedSessions.map((session, index) => (
-        <View key={index} style={styles.sessionContainer}>
-          <View style={styles.sessionBlock}>
-            <Text style={styles.sessionText}>
-             u/{session.postAuthor ? session.postAuthor : 'No name'}
-            </Text>
-          </View>
-          <View style={styles.sessionBlock}>
-            {/* <Text style={styles.sessionTitle}>Topic:</Text> */}
-            <Text style={{fontWeight:'bold', fontSize: 20}}>{session.postTopic}</Text>
-          </View>
-          <View style={styles.sessionBlock}>
-            {/* <Text style={styles.sessionTitle}>Post content:</Text> */}
-            <Text style={styles.sessionText}>
-              {session.postContent ? session.postContent : 'No content'}
-            </Text>
-          </View>
-          <TouchableOpacity
-              style={styles.deleteButton}
-              onPress={() => deleteSession(session.postId, session.userId)} // Call deleteSession function with session ID
-            >
-              <Text style={styles.deleteButtonText}>Delete</Text>
-            </TouchableOpacity>
-
-        </View>
-        
-      ))}
-    </View>
+        {sessions.map((session, index) => (
+            <View style={styles.sessionContainer}>
+              <View style={styles.sessionBlock}>
+                <Text style={styles.sessionText}>u/{session.postAuthor ? session.postAuthor : 'No name'}</Text>
+              </View>
+              <TouchableOpacity
+            key={index}
+            onPress={() => navigation.navigate(ROUTES.POST_INFORMATION, { postId: session.id })}
+          >
+              <View style={styles.sessionBlock}>
+              <Text style={{fontWeight:'bold', fontSize: 18}}>{session.postTopic}</Text>
+              </View>
+              <View style={styles.sessionBlock}>
+                <Text style={styles.sessionText}>{session.postContent ? session.postContent : 'No content'}</Text>
+              </View>
+              </TouchableOpacity>
+              <TouchableOpacity
+                style={styles.deleteButton}
+                onPress={() => deleteSession(session.postId, session.userId)}
+              >
+                <Text style={styles.deleteButtonText}>Delete</Text>
+              </TouchableOpacity>
+            </View>
+        ))}
+      </View>
     </ScrollView>
   );
 };
 
-export default CommunityScreen;
-
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    padding: 20,
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-    marginBottom: 20,
+    padding: 5,  //can be used effficiently to adjust size of the posts
   },
   sessionContainer: {
     marginBottom: 20,
@@ -106,8 +86,6 @@ const styles = StyleSheet.create({
     borderColor: '#ccc',
     padding: 10,
     borderRadius: 5,
-    width: '110%',    //change the sessioncontainer width because finished time was cut off
-    marginLeft: -10,  //change the marginleft because finished time was cut off
   },
   sessionBlock: {
     flexDirection: 'row',
@@ -140,7 +118,6 @@ const styles = StyleSheet.create({
     borderRadius: 5,
     alignItems: 'center',
   },
-  buttonContainer: {
-    paddingHorizontal: 16,
-  },
 });
+
+export default CommunityScreen;
